Spring security part2 completed

+++ src/main/java/com/project1/journalApp/service/JournalEntryService.java
@@ -47,11 +47,20 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public String DeleteById(ObjectId id, String userName){
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveNewUser(user);
-        journalEntryRepository.deleteById(id);
-        return "Data deleted";
+    public boolean DeleteById(ObjectId id, String userName){
+        boolean removed = false;
+        try {
+            User user = userService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if (removed) {
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+        } catch (Exception e) {
+            System.out.println(e);
+            throw new RuntimeException(" An error occured while deleting user !! ");
+        }
+
+        return removed;
     }
 }
